+++ libraries/urlUtils/urlUtils.js
@@ -1,0 +1,7 @@
+export function tryAppendQueryString(existingUrl, key, value) {
+  if (value) {
+    return existingUrl + key + '=' + encodeURIComponent(value) + '&';
+  }
+
+  return existingUrl;
+}
