+++ modules/adWMGBidAdapter.js
@@ -1,9 +1,9 @@
 'use strict';
 
-import { tryAppendQueryString } from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { config } from '../src/config.js';
 import { BANNER } from '../src/mediaTypes.js';
+import {tryAppendQueryString} from '../libraries/urlUtils/urlUtils.js';
 
 const BIDDER_CODE = 'adWMG';
 const ENDPOINT = 'https://hb.adwmg.com/hb';
