+++ modules/open8BidAdapter.js
@@ -1,9 +1,10 @@
 import { Renderer } from '../src/Renderer.js';
 import {ajax} from '../src/ajax.js';
-import { createTrackPixelHtml, logError, logWarn, tryAppendQueryString } from '../src/utils.js';
+import { createTrackPixelHtml, logError, logWarn} from '../src/utils.js';
 import { registerBidder } from '../src/adapters/bidderFactory.js';
 import { VIDEO, BANNER } from '../src/mediaTypes.js';
 import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
+import {tryAppendQueryString} from '../libraries/urlUtils/urlUtils.js';
 
 const BIDDER_CODE = 'open8';
 const URL = 'https://as.vt.open8.com/v1/control/prebid';
