+++ modules/trionBidAdapter.js
@@ -1,7 +1,8 @@
-import { parseSizesInput, tryAppendQueryString } from '../src/utils.js';
+import { parseSizesInput} from '../src/utils.js';
 import {registerBidder} from '../src/adapters/bidderFactory.js';
 import { getStorageManager } from '../src/storageManager.js';
 import {getBidIdParameter} from '../libraries/bidderUtils/bidderUtils.js';
+import {tryAppendQueryString} from '../libraries/urlUtils/urlUtils.js';
 
 const BID_REQUEST_BASE_URL = 'https://in-appadvertising.com/api/bidRequest';
 const USER_SYNC_URL = 'https://in-appadvertising.com/api/userSync.html';
